pyproject.toml
[tool.poetry]
name = "chatapp"
version = "0.1.0"
description = "for chatapp configuration"
authors = ["test user"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12"
bcrypt = "4.0.1"
black = "23.3.0"
django = "4.2.1"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.2.2"
flake8 = "6.0.0"
mypy = "1.3.0"
mysqlclient = "2.1.1"
isort = "5.12.0"
pytest-django = "4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "chatapp.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
follow_imports = "skip"
show_column_numbers = true
strict_optional = true
warn_no_return = true
exclude = "settings.py|migrations|tests"
